+++ source/extensions/filters/network/zookeeper_proxy/filter.cc
@@ -1,4 +1,4 @@
-#include "extensions/filters/network/zookeeper_proxy/zookeeper_filter.h"
+#include "extensions/filters/network/zookeeper_proxy/filter.h"
 
 #include <string>
 #include <vector>
