+++ source/extensions/filters/network/zookeeper_proxy/utils.cc
@@ -1,4 +1,4 @@
-#include "extensions/filters/network/zookeeper_proxy/zookeeper_utils.h"
+#include "extensions/filters/network/zookeeper_proxy/utils.h"
 
 #include <string>
